Fix some refactoring that was bad
