+++ src/components/macfast/save-for-later.tsx
@@ -1,9 +1,9 @@
-import { setSavedForLater, setSavedForLaterDebounced } from "@/lib/api";
+import { setSavedForLater } from "@/lib/question-api";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
 import { useAuthFetch } from "@/hooks/useFetchWithAuth";
 import { useEffect, useState } from "react";
-import { debounce, set } from "lodash";
+import { debounce } from "lodash";
 
 interface SaveForLaterProps {
   courseCode: string | null;
